+++ lib/hangout-farming.js
@@ -6,29 +6,29 @@ var logger = require("./logging").getLogger();
 
 var oauth2Client;
 
-var curApp;
+var curDb;
 
 var numHangoutUrlsAvailable = 0;
 
 // The rationale for this crazy situation is covered in DEVELOPMENT.md
 
-exports.init = function(app) {
-	curApp = app;
+exports.init = function(app, db, options) {
+	curDb = db;
 	
 	// keep track of the total number of hangout urls we've farmed so far, so we
 	// can put it in the UI.
-	app.redis.llen('global:hangout_urls', function(err, len) {
+	db.redis.llen('global:hangout_urls', function(err, len) {
 		numHangoutUrlsAvailable = len;
 	});
 
 	// set up an HTTP endpoint that will put us in the endpoint redirect loop
 	// people who load this url 
-	app.express.get('/hangout-farming', function(req, res){
-		var url = req.protocol + "://" + app.options.HOST + ":" + app.options.PORT + "/hangout-callback";
+	app.get('/hangout-farming', function(req, res){
+		var url = req.protocol + "://" + options.HOST + ":" + options.PORT + "/hangout-callback";
 
 		// initialize an oauth client request, using our google client id
 		oauth2Client =
-		new OAuth2Client(app.options.GOOGLE_CLIENT_ID, app.options.GOOGLE_CLIENT_SECRET, url);
+		new OAuth2Client(options.GOOGLE_CLIENT_ID, options.GOOGLE_CLIENT_SECRET, url);
 
 		var url = oauth2Client.generateAuthUrl({
 			access_type: 'offline',
@@ -43,7 +43,7 @@ exports.init = function(app) {
 	// from google, after they authorize our app. The request they make back to
 	// our server will have a token attached to it, that we will use to 
 	// place requests on behalf of this user.
-	app.express.get('/hangout-callback', function(req, res) {
+	app.get('/hangout-callback', function(req, res) {
 		// console.log(req.query["code"]);
 		oauth2Client.getToken(req.query["code"], function(err, token) {
 			if(err) {
@@ -88,10 +88,10 @@ exports.init = function(app) {
 						// event object.
 						if(!err && "hangoutLink" in event) {
 							// push it into redis for safekeeping. 
-							app.redis.lpush("global:hangout_urls", event.hangoutLink, function(err, num) {
+							db.redis.lpush("global:hangout_urls", event.hangoutLink, function(err, num) {
 								logger.info("logged new hangout url: " + event.hangoutLink + "; total: " + num);
 								// purely for convenience of clicking to farm another url.
-								res.send("urls available: " + num + "<br><a href='//" + app.options.HOST + ":" + app.options.PORT + "/hangout-farming'>CLICK ME</a>");
+								res.send("urls available: " + num + "<br><a href='//" + options.HOST + ":" + options.PORT + "/hangout-farming'>CLICK ME</a>");
 								
 								// update the tracking number.
 								numHangoutUrlsAvailable = num;
@@ -112,7 +112,7 @@ exports.init = function(app) {
 								// we send this error if the calendar event created doesn't have a hangout url
 								// included in the object. Give the user some rudimentary instructions about
 								// how to fix this situation.
-								res.send("your account doesn't seem to have the 'create video calls for events I create' option enabled. go to <a href='http://calendar.google.com'>google calendar</a>, settings (in the upper right hand corner) and enable that option. Then <a href='http://" + app.options.HOST + ":" + app.options.PORT + "/hangout-farming'>CLICK HERE!</a><br>" + JSON.stringify(event));
+								res.send("your account doesn't seem to have the 'create video calls for events I create' option enabled. go to <a href='http://calendar.google.com'>google calendar</a>, settings (in the upper right hand corner) and enable that option. Then <a href='http://" + options.HOST + ":" + options.PORT + "/hangout-farming'>CLICK HERE!</a><br>" + JSON.stringify(event));
 							}
 						}
 					});
@@ -128,7 +128,7 @@ exports.init = function(app) {
 exports.getNextHangoutUrl = function(callback) {
 	
 	// just hit redis directly here, with an rpop.
-	curApp.redis.rpop("global:hangout_urls", function(err, url) {
+	curDb.redis.rpop("global:hangout_urls", function(err, url) {
 		logger.debug("returning hangout url: " + url + " (err: " + err + ")");
 		callback && callback(err, url);
 	});	
@@ -138,7 +138,7 @@ exports.getNextHangoutUrl = function(callback) {
 // (this situation occurs in some limited situations where people
 //  join a new hangout in rapid succession)
 exports.reuseUrl = function(url) {
-	curApp.redis.rpush("global:hangout_urls", url, function(err) {
+	curDb.redis.rpush("global:hangout_urls", url, function(err) {
 		logger.debug("re-added a url that was no longer needed: " + url);
 	});
 }
