+++ lib/passport-mock.js
@@ -39,7 +39,7 @@ var mockUsers = {
 
 exports.mockAuthMiddleware = function(server) {
     _.each(mockUsers, function(user) {
-        server.users.add(new models.ServerUser(user));
+        server.db.users.add(new models.ServerUser(user));
     });
 
     function mockAuth(req, res, next) {
@@ -52,7 +52,7 @@ exports.mockAuthMiddleware = function(server) {
             // unauthenticated.
             return next();
         }
-        req.user = server.users.findWhere({"sock-key": mock_user});
+        req.user = server.db.users.findWhere({"sock-key": mock_user});
         next();
     };
     return mockAuth;
